requirements.txt
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
boto3==1.23.3
botocore==1.26.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
comtypes==1.1.11
customtkinter==3.11
darkdetect==0.5.1
docopt==0.6.2
filelock==3.7.0
huggingface-hub==0.7.0
idna==3.3
jmespath==1.0.0
joblib==1.1.0
Js2Py==0.71
nltk==3.7
numpy==1.22.3
packaging==21.3
Pillow==9.1.1
pipwin==0.5.2
protobuf==4.21.0
pyjsparser==2.7.1
pyparsing==3.0.9
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
pyttsx3==2.90
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
s3transfer==0.5.2
sacremoses==0.0.53
sentencepiece==0.1.96
six==1.16.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.19.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
